+++ CHANGELOG.md
@@ -7,10 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## NOTE
 
+- after update to 5.5.x the volume control need to be configured again
 - after update to 4.9.x all buttons (Sound Mode) need to be configure again using config UI
 - after update to 4.8.x all buttons need to be configure again using config UI
 - after update to 4.7.x buttons, sensors, volume display type need to be configure again using config UI
 - after update to 3.15.x need remove the accessory frome Home app and add it again
+
+## [5.5.0] - (16.02.2025)
+
+## Changes
+
+- after update to this version the volume control need to be configured again
+- fix [#303](https://github.com/grzegorz914/homebridge-denon-tv/issues/303)
+- fix [#305](https://github.com/grzegorz914/homebridge-denon-tv/issues/305)
+- cleanup
 
 ## [5.4.5] - (09.02.2025)